Fix autostart, make code more robust

+++ swm/src/autostart.c
@@ -20,8 +20,8 @@ void Autostart__add(Autostart *as, const char *cmd)
 
 	if (as->size < MAX_AUTOSTART_COMMANDS + 1) {
 		as->cmds[as->size++] = (char *)cmd;
+		as->pids[as->size] = -1;
 		as->size += 1;
-		as->cmds[as->size++] = NULL;
 	}
 }
 
@@ -31,7 +31,6 @@ void Autostart__exec(Autostart *as)
 	if (!as) {
 		return;
 	}
-
 	char buffer[MAX_AUTOSTART_COMMANDS] = { 0 };
 
 	for (uint32_t i = 0; i < as->size; ++i) {
@@ -40,9 +39,13 @@ void Autostart__exec(Autostart *as)
 		} else {
 			if (buffer[0] != '\0') {
 				if (system_exec(buffer, &as->pids[i]) < 0) {
+					fprintf(stderr,
+						"'%s' failed to start\n",
+						buffer);
 					as->pids[i] = -1;
 				}
-				// reset command buffer
+
+				// We need to reset the buffer for the next command to be loaded in.
 				memset(buffer, 0, sizeof(buffer));
 				buffer[0] = '\0';
 			}
@@ -59,7 +62,11 @@ void Autostart__kill(Autostart *as)
 	}
 
 	for (uint32_t i = 0; i < as->size; ++i) {
-		if (as->pids[i] > 0) {
+		// The `as->size` variable contains null entries to seperate
+		// commands. The pid array will hold no such null seperator values.
+		// Basically ever odd index of the pid array will hold a real pid,
+		// so if pid % 2 is 1, we have an odd index
+		if ((as->pids[i] % 2 == 1) && as->pids[i] > 0) {
 			kill(as->pids[i], SIGTERM);
 			waitpid(as->pids[i], NULL, WNOHANG);
 		}
